bash-completion: use modern lazy-load directory
The legacy bash_completion.d directory requires every script be sourced
at shell startup, which is slow. Since bash-completion 2.x (released in
2012), completions are loaded on-demand -- a minimal completion stub is
defined as the default command completer, which will search for a script
in the completionsdir named the same as the command, and source this
script in order to update the command completion.

Make use of this lazy-loaded completion. This entails:
- moving the installation location
- getting rid of the legacy modification of .bashrc which is no longer
  needed